Change threads structure
In the previuos version, App.state.threads is a list, which is hard
to add or find a target. In this commit, threads is changed to a
dictionary with id_ as key. 'Public Room' and 'Miss Google' are
two default id_s. New id_s could be added in the future for private
talks.